remove unused props
