+++ src/components/filters/filters.js
@@ -11,8 +11,6 @@ const Filters = ({
   onGroupToggleChanged,
   onItemChange,
   onItemClick,
-  onItemMouseEnter,
-  onItemMouseLeave,
   onResetFilter,
   onToggleGroupCollapsed,
   searchValue,
@@ -42,8 +40,6 @@ const Filters = ({
               onGroupToggleChanged={onGroupToggleChanged}
               onItemChange={onItemChange}
               onItemClick={onItemClick}
-              onItemMouseEnter={onItemMouseEnter}
-              onItemMouseLeave={onItemMouseLeave}
               onToggleGroupCollapsed={onToggleGroupCollapsed}
               searchValue={searchValue}
             />
